server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cloudinary==1.44.2
distro==1.9.0
fastapi==0.136.1
flatbuffers==25.12.19
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
ImageIO==2.37.3
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
llvmlite==0.47.0
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
onnxruntime==1.25.1
openai==2.32.0
packaging==26.2
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
protobuf==7.34.1
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
PyMatting==1.1.15
python-dotenv==1.2.2
python-multipart==0.0.26
referencing==0.37.0
rembg==2.0.75
requests==2.33.1
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
tifffile==2026.4.11
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
